+++ app/controllers/projects_controller.rb
@@ -134,7 +134,7 @@ class ProjectsController < ApplicationController
 
   def project_params
     params.require(:project).permit(
-      :name, :slug, :description, :subject_code_name, :show_contacts,
+      :name, :slug, :description, :subject_code_name,
       :show_documents, :disable_all_emails,
       :collect_email_on_surveys, :hide_values_on_pdfs,
       :randomizations_enabled, :adverse_events_enabled, :blinding_enabled,