+++ app/models/project.rb
@@ -64,7 +64,6 @@ class Project < ActiveRecord::Base
   has_many :events, -> { where deleted: false }
   has_many :handoffs
 
-  has_many :contacts, -> { where deleted: false }
   has_many :documents, -> { where deleted: false }
 
   has_many :domains, -> { where deleted: false }